requirements.txt
typer[all]>=0.12.0
pydantic>=2.10.0
pydantic-settings>=2.3.0
sqlalchemy>=2.0.0
yoyo-migrations>=7.3.0
pandas>=2.2.0
numpy>=2.0.0
requests>=2.32.0
python-dotenv>=1.0.0
streamlit>=1.57.0
plotly>=5.24.0
ollama>=0.6.2
apscheduler>=3.11.1,<4.0
psycopg2-binary>=2.9.10
pyarrow>=24.0.0
openpyxl>=3.1.0
